fix
The remaining blocker was a signed/unsigned lane comparison in the shared hash-table code, not the test file itself. I patched the leader lane index to an unsigned type at the call sites in single_value_hash_table.hpp, hash_set.hpp, multi_value_hash_table.hpp, multi_bucket_hash_table.hpp, and gpu_engine.hpp.

+++ include/warpcore/gpu_engine.hpp
@@ -77,7 +77,8 @@ void for_each_unique_key(
 
                 if(group.any(hit))
                 {
-                    const auto leader = detail::first_set_bit(hit_mask);
+                    const auto leader =
+                        static_cast<unsigned int>(detail::first_set_bit(hit_mask));
 
                     // check if search_key is the first entry for this key
                     if(group.thread_rank() == leader && i == gid)
